unsecure/backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
mysql-connector-python